% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{analyse_fisher}
\alias{analyse_fisher}
\title{analyse_fisher}
\usage{
analyse_fisher(datalist, sens_status_predicted)
}
\arguments{
\item{datalist}{sens_status_predicted}
}
\value{
pval P-value from Fisher's exact test
}
\description{
Perform Fisher's exact test for the sensitive group
}
\author{
Svetlana Cherlin, James Wason
}
