% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseRealData.R
\name{analyse_realdata}
\alias{analyse_realdata}
\title{Analysis of the the real data (retrospective) according to the cross-validate adaptive enrichment risk scores (CADEN) design}
\usage{
analyse_realdata(
  datalist_stage1,
  datalist_stage2,
  threshold_overall,
  threshold_group,
  seed,
  standardise_cvrs = TRUE,
  full_model = TRUE
)
}
\arguments{
\item{threshold_overall}{P-value threshold for the test for the differences in the treatment effect in the overall trial population}

\item{threshold_group}{P-value threshold for the test for the treatment effect in the sensitive group}

\item{seed}{A seed for random number generating}

\item{standardise_cvrs}{A logical flag for the standardisation of the risk scores. Default is 'standardise_cvrs=TRUE'.
The standardisation is performed with respect to the training data sets, per cross-validation fold.}

\item{full_model}{A logical flag for the full model (treatment effect, covariate effect and the interaction effect). Default is 1full_model = TRUE. When 'full_model = FALSE', only interaction effect is included in the model}

\item{realdata_stage1}{A list with two data frames (patients, covar) and two vectors (resp.rate, response)
patients: a data frame with one row per patient and the following columns:
FID (family ID), IID (individual ID), treat (1 for treatment and 0 for control),
sens_status (true sensitivity status), stage (1)
covar: a data frame with covariate data
response: a vector of simulated binary responses}

\item{realdata_stage2}{A list with two data frames with the same structure as realdata_stage1}
}
\value{
An object of class \code{"caden"}.

decision: Enrichment, stop or unselected.

cvrs: A vector of the risk scores

sens_status_predicted: A vector of the predicted sensitivity status.

noneligible: Number of patients non-eligible for the trial (for the enrichment strategy).

pval_overall: P-value for the difference between the arm in the overall trial population.

pval_fisher_group: P-value for the difference between the arms in the sensitive group (as assessed using Fisher exact test).
}
\description{
The design has two stages. At the end of stage 1, an interim analysis is performed to test the efficacy of the treatment compared to the control in the overall trial population. If the test in the overall population is significanct, the trial proceeds into stage 2 by analysing all patients (the "unselected" strategy). If the test in the overall population is not significant then the trial proceeds by testing whether there is a subgroup of patients (sensitive group) that show a promising treatment effect. Depending on the results of the test in the sensitive group, the design proceeds into stage 2 according to one of the mutually exclusive strategies "stop" or "enrichment".
}
\examples{
data(realdata_stage1)
data(realdata_stage2)
threshold_overall = 0.04
threshold_group = 0.1
seed = 123
standardise_cvrs = 0
full_model = 0
real_res <- analyse_realdata(realdata_stage1, realdata_stage2, threshold_overall, threshold_group, seed, standardise_cvrs, full_model)
}
\seealso{
\code{\link{simulate_data}}, function.
}
\author{
Svetlana Cherlin, James Wason
}
