% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseSimData.R
\name{analyse_simdata}
\alias{analyse_simdata}
\title{Analysis of the simulated data according to the cross-validate adaptive enrichment risk scores (CADEN) design}
\usage{
analyse_simdata(param_file, datalist_stage1)
}
\arguments{
\item{param_txt}{A name of the text file with the parameters as described in simulate.R}

\item{simdata_stage1}{A list with the simulated data according to the output of simulate.R}
}
\value{
An object of class \code{"caden"}.

datalist_stage2: a list of simulated data for stage 2

sens_status_predicted: A vector of the predicted sensitivity status

sensitivity: A vector of values for sensitivity for identifying the sensitive group

specificity: A vector of values for specificity for identifying the sensitive group

noneligible: Number of patients non-eligible for the trial (for the enrichment strategy).

pval_overall: P-value for the difference between the arm in the overall trial population (for "unselected" strategy only)

pval_fisher_group: P-value for the difference between the arms in the sensitive group (as assessed using Fisher exact test)
}
\description{
The design has two stages. At the end of stage 1, an interim analysis is performed to test the efficacy of the treatment compared to the control in the overall trial population. If the test in the overall population is significanct, the trial proceeds into stage 2 by enrolling all patients (the "unselected" strategy) implemented here via simulating the data for stage 2 patients. If the test in the overall population is not significant then the trial proceeds by testing whether there is a subgroup of patients (sensitive group) that show a promising treatment effect. Depending on the results of the test in the sensitive group, the design proceeds into stage 2 according to one of the mutually exclusive strategies "stop" or "enrichment".
}
\examples{
data(simdata_stage1)
param_file <- "data/param.txt"
sim_res <- analyse_simdata(param_file, simdata_stage1)
}
\seealso{
\code{\link{simulate_data}}, function.
}
\author{
Svetlana Cherlin, James Wason
}
