% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{analyse_stage1}
\alias{analyse_stage1}
\title{analyse_stage1}
\usage{
analyse_stage1(
  datalist_stage1,
  threshold_overall,
  threshold_group,
  seed,
  standardise_cvrs,
  full_model
)
}
\arguments{
\item{datalist_stage1}{threshold_overall
threshold_group
standardise_cvrs
full_model}
}
\value{
A list with the following elements
         decision
         sens_status_predicted
         model
}
\description{
Perform interim analysis
}
\author{
Svetlana Cherlin, James Wason
}
