% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{analyse_stage2}
\alias{analyse_stage2}
\title{analyse_stage2}
\usage{
analyse_stage2(decision, params, model, standardise_cvrs)
}
\arguments{
\item{decision}{params
model
standardise_cvrs}
}
\value{
A list with the following elements
         datalist_stage2
         noneligible
         sensitivity
         specificity
         sens_status_predicted
}
\description{
Analyse the stage 2 of the trial
}
\author{
Svetlana Cherlin, James Wason
}
