% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{analyse_subgroup}
\alias{analyse_subgroup}
\title{analyse_subgroup}
\usage{
analyse_subgroup(datalist, sens_status_predicted)
}
\arguments{
\item{datalist}{sens_status_predicted}
}
\value{
pval_treat_group P-value for the treatment effect in the sensitive group
}
\description{
Compute the treatment effect in the sensitive group
}
\author{
Svetlana Cherlin, James Wason
}
