% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check_eligibility}
\alias{check_eligibility}
\title{}
\usage{
check_eligibility(covar, model, standardise_cvrs)
}
\arguments{
\item{covar}{model
standardise_cvrs}
}
\value{
A list with the following elements
         eligible
         sens_status_predicted
         cvrs
}
\description{
Compute the eligibility status for stage 2
}
\author{
Svetlana Cherlin, James Wason
}
