% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStage2.R
\name{get_stage2}
\alias{get_stage2}
\title{Simulate a data set for stage 2 for the the cross-validate adaptive enrichment risk scores (CADEN) design}
\usage{
get_stage2(datalist1, seed)
}
\arguments{
\item{datalist1}{A list with two data frames (patients, covar) and avector of binary responses
patients: a data frame with one row per patient and the following columns:
FID (family ID), IID (individual ID), treat (1 for treatment and 0 for control)
covar: a data frame with covariate data
response: a vector of simulated binary responses}

\item{seed}{}
}
\value{
datalist2 A list with two data frames (patients, covar) and avector of binary responses
        patients: a data frame with one row per patient and the following columns:
        FID (family ID), IID (individual ID), treat (1 for treatment and 0 for control)
        covar: a data frame with covariate data
        response: a vector of simulated binary responses
}
\description{
Create a data set for stage 2 by sampling with replacement from the stage 1 data set
}
\examples{
data(realdata_stage1)
data(realdata_stage2)
real <- analyseReal(realdata_stage1, realdata_stage2, 0.04, 0.01, 123)
}
\seealso{
\code{\link{analyse_realdata}}, function.
}
\author{
Svetlana Cherlin, James Wason
}
