% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata_stage1-data.R
\docType{data}
\name{simdata_stage1}
\alias{simdata_stage1}
\title{Simulated data}
\format{
A list with two data frames (patients, covar) and two vectors (resp.rate, response)
        A list with two data frames (patients, covar) and two vectors (resp.rate, response)
        patients: a data frame with one row per patient and the following columns:
        FID (family ID), IID (individual ID), treat (1 for treatment and 0 for control),
        sens.status (true sensitivity status), stage (1)
        covar: a data frame with covariate data
        resp_rate: a vector of response rates
        response: a vector of simulated binary responses
}
\usage{
data(simdata_stage1)
}
\description{
Simulated data for the cross-validate adaptive enrichment risk scores (CADEN) design
}
\examples{

data(simdata_stage1)
\donttest{
simres <- analyse_simdata("param.file", simdata_stage1)
}

}
\keyword{datasets}
